Split app files
I split the app.py file into modules for app, db, and user profile
