+++ app.py
@@ -1,8 +1,9 @@
 from flask import Flask, request, render_template, redirect, url_for, flash
-import mysql.connector
 import stripe
 import os
 from dotenv import load_dotenv
+from user_routes import user_bp
+from db import get_db_connection
 
 # Load environment variables from .env file (optional but helpful for dev)
 load_dotenv()
@@ -10,22 +11,10 @@ load_dotenv()
 # Flask app setup
 app = Flask(__name__)
 app.secret_key = os.getenv('FLASK_SECRET_KEY', 'supersecret')  # Needed for flash messages
-
-# MySQL configuration from environment
-db_config = {
-    'host': os.getenv('DB_HOST', 'localhost'),
-    'user': os.getenv('DB_USER', 'your_username'),
-    'password': os.getenv('DB_PASS', 'your_password'),
-    'database': os.getenv('DB_NAME', 'artisan_platform'),
-    'ssl_ca': os.getenv('SSL_CA')
-}
+app.register_blueprint(user_bp)
 
 # Stripe Test Mode key
 stripe.api_key = os.getenv('STRIPE_SECRET_KEY', 'sk_test_your_test_key')
-
-# Helper function to connect to MySQL
-def get_db_connection():
-    return mysql.connector.connect(**db_config)
 
 # Routes
 
@@ -147,58 +136,6 @@ def subscribe(artisan_id):
     conn.close()
 
     return render_template('subscribe.html', artisan=artisan)
-
-@app.route('/profile', methods=['GET'])
-def profile():
-    user_id = request.args.get('user_id')
-    if not user_id:
-        flash("User ID not provided.", "danger")
-        return redirect(url_for('index'))
-
-    conn = get_db_connection()
-    cursor = conn.cursor(dictionary=True)
-    cursor.execute("SELECT * FROM users WHERE id = %s", (user_id,))
-    user = cursor.fetchone()
-    cursor.close()
-    conn.close()
-
-    if not user:
-        flash("User not found.", "danger")
-        return redirect(url_for('index'))
-
-    return render_template('profile.html', user=user)
-
-
-@app.route('/update_profile', methods=['POST'])
-def update_profile():
-    user_id = request.form.get('user_id')
-    email = request.form.get('email')
-    first_name = request.form.get('first_name')
-    last_name = request.form.get('last_name')
-    phone = request.form.get('phone')
-    address_line1 = request.form.get('address_line1')
-    city = request.form.get('city')
-    state = request.form.get('state')
-    zip_code = request.form.get('zip_code')
-
-    if not user_id or not email or not first_name or not last_name:
-        flash("First name, last name, and email are required.", "danger")
-        return redirect(url_for('profile', user_id=user_id))
-
-    conn = get_db_connection()
-    cursor = conn.cursor()
-    cursor.execute("""
-        UPDATE users
-        SET email=%s, first_name=%s, last_name=%s, phone=%s,
-            address_line1=%s, city=%s, state=%s, zip_code=%s
-        WHERE id=%s
-    """, (email, first_name, last_name, phone, address_line1, city, state, zip_code, user_id))
-    conn.commit()
-    cursor.close()
-    conn.close()
-
-    flash("Profile updated successfully.", "success")
-    return redirect(url_for('profile', user_id=user_id))
 
 # Run the app
 if __name__ == '__main__':
